Avoid any usage of ssize_t

+++ connection.c
@@ -700,7 +700,7 @@ dl_read (DLCP *dlconn, int64_t pktid, DLPacket *packet, void *packetdata,
     packet->datasize  = sdatasize;
 
     /* Check that the packet data size is not beyond the max receive buffer size */
-    if (packet->datasize > (ssize_t)maxdatasize)
+    if (packet->datasize > (int64_t)maxdatasize)
     {
       dl_log_r (dlconn, 2, 0,
                 "[%s] dl_read(): packet data larger (%d) than receiving buffer (%" PRIsize_t ")\n",
@@ -858,7 +858,7 @@ dl_getinfo (DLCP *dlconn, const char *infotype, char *infomatch,
     }
 
     /* If a maximum buffer size was specified check that it's large enough */
-    if (maxinfosize && infosize > (ssize_t)maxinfosize)
+    if (maxinfosize && infosize > (int64_t)maxinfosize)
     {
       dl_log_r (dlconn, 2, 0, "[%s] dl_getinfo(): INFO data larger (%d) than the maximum size (%" PRIsize_t ")\n",
                 dlconn->addr, infosize, maxinfosize);
@@ -1071,7 +1071,7 @@ dl_collect (DLCP *dlconn, DLPacket *packet, void *packetdata,
           packet->dataend   = sdataend;
           packet->datasize  = sdatasize;
 
-          if (packet->datasize > (ssize_t)maxdatasize)
+          if (packet->datasize > (int64_t)maxdatasize)
           {
             dl_log_r (dlconn, 2, 0,
                       "[%s] dl_collect(): packet data larger (%d) than receiving buffer (%" PRIsize_t ")\n",
@@ -1289,7 +1289,7 @@ dl_collect_nb (DLCP *dlconn, DLPacket *packet, void *packetdata,
       packet->dataend   = sdataend;
       packet->datasize  = sdatasize;
 
-      if (packet->datasize > (ssize_t)maxdatasize)
+      if (packet->datasize > (int64_t)maxdatasize)
       {
         dl_log_r (dlconn, 2, 0,
                   "[%s] dl_collect_nb(): packet data larger (%d) than receiving buffer (%" PRIsize_t ")\n",
